GithubCommand: Allow specifiying a different repository
This patch adds the option for the user to specify the repository they
want to view a pull request or issue of.

If no repository was specified explicitly, the category the command was
executed in will be used to infer a repository.

If the category does not have a repository associated with it, the
command will default to the SerenityOS/serenity repository.

As a visual indication, embeds now show the repository name before
the issue / pull id:

`serenity #1337: Meta: Fixed something`